added userId as creator for authorization

+++ backend/routes/posts.js
@@ -4,7 +4,7 @@ const multer = require('multer');
 const Post = require("../models/post");
 const checkAuth = require("../middleware/check-auth");
 
-const router = express.Router();
+const router = express.Router(); 
 
 const MIME_TYPE_MAP ={
   'image/png': 'png',
@@ -35,7 +35,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   console.log(post);
   post.save().then(createdPost => {
@@ -62,12 +63,18 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({ _id: req.params.id}, post).then(result => {
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId}, post).then(result => {
     console.log(result);
-    res.status(200).json({message: "Update successful!"});
+    if(result.nModified > 0) {
+      res.status(200).json({ message: "Update successful!" });
+    }
+    else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
   })
 });
 
@@ -110,9 +117,14 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({message: 'Post deleted!'});
+    if(result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    }
+    else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
 
   })
